Extracts handlers to hooks
