+++ console/src/fusion-access/components/Page.tsx
@@ -1,10 +1,8 @@
-import { useCallback, useLayoutEffect } from "react";
+import { useLayoutEffect } from "react";
 import { Helmet } from "react-helmet";
 import {
   ListPageBody,
   ListPageHeader,
-  useK8sModel,
-  k8sCreate,
 } from "@openshift-console/dynamic-plugin-sdk";
 import {
   Alert,
@@ -14,23 +12,16 @@ import {
 import { useWatchSpectrumScaleCluster } from "../hooks/useWatchSpectrumScaleCluster";
 import { useWatchFusionAccess } from "../hooks/useWatchFusionAccess";
 import { useGlobalStateContext } from "../contexts/global-state/GlobalStateContext";
-import { STORAGE_ROLE_LABEL } from "../constants";
 import { DownloadLogsButton } from "./DownloadLogsButton";
 import { CreateStorageClusterButton } from "./CreateStorageClusterButton";
 import { usePageSectionRouter } from "../hooks/usePageSectionRouter";
 import { useWatchNode } from "../hooks/useWatchNode";
 import { useSelectedNodes } from "../hooks/useSelectedNodes";
 import { useTriggerAlertsOnErrors } from "../hooks/useTriggerAlertsOnErrors";
-import { usePluginTranslations } from "@/hooks/usePluginTranslations";
-import type { Cluster } from "@/models/ibm-spectrum-scale/Cluster";
-
-const [storageRoleLabelKey, storageRoleLabelValue] =
-  STORAGE_ROLE_LABEL.split("=");
-const nodeSelector = { [storageRoleLabelKey]: storageRoleLabelValue };
+import { useCreateStorageClusterHandler } from "../hooks/useCreateStorageClusterHandler";
 
 export const Page: React.FC = () => {
   const [state, dispatch] = useGlobalStateContext();
-  const { t } = usePluginTranslations();
 
   useLayoutEffect(() => {
     // Ensure the body takes full height, to allow displaying warnings at the bottom of the same section
@@ -75,48 +66,7 @@ export const Page: React.FC = () => {
     fusionAccessesListLoaded,
   });
 
-  const [storageScaleClusterModel] = useK8sModel({
-    group: "scale.spectrum.ibm.com",
-    version: "v1beta1",
-    kind: "Cluster",
-  });
-
-  const handleCreateStorageCluster = useCallback(
-    async () => {
-      try {
-        await k8sCreate<Cluster>({
-          model: storageScaleClusterModel,
-          data: {
-            apiVersion: "scale.spectrum.ibm.com/v1beta1",
-            kind: "Cluster",
-            metadata: { name: "ibm-spectrum-scale-cluster2" },
-            spec: {
-              license: { accept: true, license: "data-management" },
-              pmcollector: {
-                nodeSelector,
-              },
-              daemon: {
-                nodeSelector,
-              },
-            },
-          },
-        });
-      } catch (e) {
-        dispatch({
-          type: "addAlert",
-          payload: {
-            key: Date.now(),
-            variant: "danger",
-            title: t("An error occurred while creating resources "),
-            description: (e as Error).message,
-            dismissable: true,
-          },
-        });
-      }
-    },
-    // Safe to ignore: 't', 'dispatch' and 'storageScaleClusterModel'
-    []
-  );
+  const handleCreateStorageCluster = useCreateStorageClusterHandler();
 
   return (
     <>
